panel1.0/backend/requirements.txt
Flask>=3.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
anthropic>=0.39.0,<1.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
tensorflow>=2.13.0
tf-keras>=2.15.0
gunicorn>=21.2.0
scikit-learn>=1.3.0
